Revert "Feature/UI (#165)" (#174)
This reverts commit c8e5d8f1745eb3659da125ac42150040cfadb171.

+++ WASD/Assets/Scripts/Player/PlayerHealthManager.cs
@@ -46,7 +46,7 @@ public class PlayerHealthManager : MonoBehaviour
         //lerpTimer = 0f;
     }
 
-    public void DamagePlayer(float damage)
+    public void DamagePlayer(int damage)
     {
         if((currentHealth - damage) <= 0)
         {
